+++ i18n/az.ts
@@ -669,7 +669,7 @@ export default {
     collaborator: "Əməkdaş səhifəsi",
     collaboratorpost: "Əməkdaş postu",
     sidebar_Home: "Ana səhifə",
-    sidebar_Page: "Səhifə",
+    sidebar_Page: "Səhifə idarəsi",
     sidebar_Message: "Mesajlar",
     sidebar_Wallet: "Cüzdan",
     sidebar_biolink: "Bio Link",